requirements.txt
asgiref==3.8.1
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
inflection==0.5.1
packaging==24.2
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
gunicorn==22.0.0
drf-yasg==1.21.7
djangorestframework_simplejwt==5.4.0
kombu==5.5.3
requests==2.32.3
openai==1.59.6
pillow==11.3.0
redis==5.2.1
celery==5.5.2
django-celery-beat==2.8.0
django_celery_results==2.6.0
kombu==5.5.3